+++ CotEditor/Sources/SnippetKeyBindingManager.swift
@@ -46,8 +46,7 @@ final class SnippetKeyBindingManager: KeyBindingManager {
     
     override private init() {
         
-        _defaultKeyBindings = [KeyBinding(action: Self.action(index: 0),
-                                          shortcut: Shortcut(modifierMask: .shift, keyEquivalent: "\r"))]
+        _defaultKeyBindings = []
         self.defaultSnippets = UserDefaults.standard.registeredValue(for: .insertCustomTextArray)
         
         super.init()
@@ -81,7 +80,7 @@ final class SnippetKeyBindingManager: KeyBindingManager {
         
         return (0..<count).map { index in
             let title = String(format: "Insert Text %li".localized, index)
-            let action = Self.action(index: index)
+            let action = self.action(index: index)
             let keyBinding = keyBindings.first { $0.action == action }
             
             let item = KeyBindingItem(action: action, shortcut: keyBinding?.shortcut, defaultShortcut: .none)
@@ -94,9 +93,7 @@ final class SnippetKeyBindingManager: KeyBindingManager {
     /// whether key bindings are not customized
     override var usesDefaultKeyBindings: Bool {
         
-        let usesDefaultSnippets = self.snippets == self.defaultSnippets
-        
-        return usesDefaultSnippets && super.usesDefaultKeyBindings
+        return (self.snippets == self.defaultSnippets) && super.usesDefaultKeyBindings
     }
     
     
@@ -128,7 +125,7 @@ final class SnippetKeyBindingManager: KeyBindingManager {
         
         guard
             let keyBinding = self.keyBindings.first(where: { $0.shortcut == shortcut }),
-            let index = Self.snippetIndex(for: keyBinding.action),
+            let index = self.snippetIndex(for: keyBinding.action),
             let snippetString = self.snippets[safe: index]
             else { return nil }
         
@@ -139,14 +136,8 @@ final class SnippetKeyBindingManager: KeyBindingManager {
     /// snippet texts to insert with key binding
     var snippets: [String] {
         
-        return UserDefaults.standard[.insertCustomTextArray] ?? []
-    }
-    
-    
-    /// save texts to insert
-    func saveSnippets(_ snippets: [String]) {
-        
-        UserDefaults.standard[.insertCustomTextArray] = snippets
+        get { UserDefaults.standard[.insertCustomTextArray] ?? [] }
+        set { UserDefaults.standard[.insertCustomTextArray] = newValue }
     }
     
     
@@ -154,25 +145,22 @@ final class SnippetKeyBindingManager: KeyBindingManager {
     // MARK: Private Methods
     
     /// build selector name for index
-    private static func action(index: Int) -> Selector {
+    private func action(index: Int) -> Selector {
         
         return Selector(String(format: "insertCustomText_%02li:", index))
     }
     
     
     /// extract index number of snippet from selector name
-    private static func snippetIndex(for action: Selector) -> Int? {
+    private func snippetIndex(for action: Selector) -> Int? {
         
-        let selectorString = NSStringFromSelector(action)
+        let selector = NSStringFromSelector(action)
         
-        guard !selectorString.isEmpty else { return nil }
+        guard
+            let range = selector.range(of: "(?<=^insertCustomText_)[0-9]{2}(?=:$)", options: .regularExpression)
+            else { return nil }
         
-        let regex = try! NSRegularExpression(pattern: "^insertCustomText_([0-9]{2}):$")
-        let result = regex.firstMatch(in: selectorString, range: selectorString.nsRange)
-        
-        guard let numberRange = result?.range(at: 1) else { return nil }
-        
-        return Int((selectorString as NSString).substring(with: numberRange))
+        return Int(selector[range])
     }
     
 }